Fix unbound variable in example.
Mathieu Lirzin reported this problem, and Per Bothner, the author of
SRFI 64, provided the patch in this change.  Here is Mathieu's error
report:

From: Mathieu Lirzin
To: srfi-64@srfi.schemers.org
Subject: Problem with 'test-group-with-cleanup' documentation.
Flags: replied, seen, list
Date: Thu 04 Aug 2016 10:26:34 AM EDT

Hello,

In the documentation the following example is given for
'test-group-with-cleanup':

  (test-group-with-cleanup "test-file"
    (define f (open-output-file "log"))
    (do-a-bunch-of-tests f)
    (close-output-port f))

However AIUI 'f' is not bound in the cleanup form since this example
expands to the following:

  (test-group suite-name
    (dynamic-wind
      (lambda ()
         #f)
      (lambda ()
        (define f (open-output-file "log"))
        (do-a-bunch-of-tests f))
      (lambda ()
        (close-output-port f))))

Would it be possible to fix the documentation by moving the definition
of 'f' to the top level?

Thanks.

--
Mathieu Lirzin